Add files via upload
This screenshot shows how after connecting to mySql and listing all the databases from previous snaphot, we can connect to a specific database and list all the tables in that database. To access a specific database "use" command is used along with the database name and after connecting to that specific database,  to list all tables "show tables;" command is used.